Moved loading state changes to a context provider
