+++ pages/[version]/[...slug].tsx
@@ -7,7 +7,7 @@ import {
 import fs from "fs";
 import matter from "gray-matter";
 import { basename } from "path";
-import { Router, useRouter } from "next/router";
+import { useRouter } from "next/router";
 import Markdoc from "@markdoc/markdoc";
 import SideNav from "../../components/SideNav/SideNav";
 import TopNav from "../../components/TopNav/TopNav";
@@ -22,10 +22,9 @@ import { getCategoryByIndex } from "../../lib/utils";
 import ErrorBoundary from "../../components/ErrorBoundary";
 import { StepsContextProvider } from "../../context/StepsContext";
 import { isEmpty, startCase } from "lodash";
-import { useDocVersionContext } from "../../context/DocVersionContext";
 import SkeletonLoader from "../../components/common/SkeletonLoader/SkeletonLoader";
-import { SkeletonWidth } from "../../enums/SkeletonLoder.enum";
 import { SKELETON_PARAGRAPH_WIDTHS } from "../../constants/SkeletonLoader.constants";
+import { useRouteChangingContext } from "../../context/RouteChangingContext";
 
 interface Props {
   menu: MenuItem[];
@@ -38,8 +37,8 @@ const SCROLLING_OFFSET = 152;
 
 export default function Article({ menu, content, slug }: Props) {
   const router = useRouter();
+  const { isRouteChanging } = useRouteChangingContext();
   const [collapsedNav, setCollapsedNav] = useState(false);
-  const [loading, setLoading] = useState(false);
   const handleCollapsedNav = (value: boolean) => {
     setCollapsedNav(value);
   };
@@ -76,23 +75,6 @@ export default function Article({ menu, content, slug }: Props) {
   }, []);
 
   useEffect(() => {
-    const onStart = () => {
-      setLoading(true);
-    };
-    const onEnd = () => {
-      setLoading(false);
-    };
-    Router.events.on("routeChangeStart", onStart);
-    Router.events.on("routeChangeComplete", onEnd);
-    Router.events.on("routeChangeError", onEnd);
-    return () => {
-      Router.events.off("routeChangeStart", onStart);
-      Router.events.off("routeChangeComplete", onEnd);
-      Router.events.off("routeChangeError", onEnd);
-    };
-  }, [Router]);
-
-  useEffect(() => {
     setCollapsedNav(isEmpty(item?.children));
   }, [item]);
 
@@ -107,11 +89,11 @@ export default function Article({ menu, content, slug }: Props) {
             category={item ? item.category : startCase(category)}
             collapsedNav={collapsedNav}
             items={item ? item.children : []}
-            loading={loading}
+            loading={isRouteChanging}
             handleCollapsedNav={handleCollapsedNav}
           />
           <main className={classNames("flex flex-col content")}>
-            {loading ? (
+            {isRouteChanging ? (
               <SkeletonLoader
                 paragraph={{
                   rows: SKELETON_PARAGRAPH_WIDTHS.length,
