+++ pages/_app.tsx
@@ -10,6 +10,7 @@ import type { MarkdocNextJsPageProps } from "@markdoc/next.js";
 import { useRouter } from "next/router";
 import ErrorBoundary from "../components/ErrorBoundary";
 import { DocVersionContextProvider } from "../context/DocVersionContext";
+import { RouteChangingContextProvider } from "../context/RouteChangingContext";
 
 const TITLE = "Markdoc";
 const DESCRIPTION = "A powerful, flexible, Markdown-based authoring framework";
@@ -43,9 +44,11 @@ export default function MyApp({ Component, pageProps }: AppProps<MyAppProps>) {
         <link rel="icon" href="/favicon.ico" />
       </Head>
       <ErrorBoundary>
-        <DocVersionContextProvider>
-          <Component {...pageProps} key={router.asPath} />
-        </DocVersionContextProvider>
+        <RouteChangingContextProvider>
+          <DocVersionContextProvider>
+            <Component {...pageProps} key={router.asPath} />
+          </DocVersionContextProvider>
+        </RouteChangingContextProvider>
       </ErrorBoundary>
     </>
   );
